feat(catalog): widen unit columns and seed the real CCS subject catalog
Phase 6 Task 1. subjects.units and enrollments.total_units were
unsignedTinyInteger/unsignedSmallInteger, which cannot represent the
real GRC catalog: Leadership subjects carry 1.5 units. Widen both to
decimal(_,1), cast as float (units is not a policy-sensitive value
like academic_grades.final_grade, so coercion is safe).

Add CcsSubjectSeeder with the 88 real CCS subjects transcribed from
two block-section spreadsheets the user supplied (kept verbatim under
docs/reference/ for provenance), resolving 13 rows where the source
spreadsheet's UNITS column actually held a SCHED ID by majority value
per subject code. Additive alongside the existing synthetic catalog --
no code collisions, all existing tests and demo lifecycles untouched.

A new integration test proves the existing Phase 5 Program Chair
curriculum/prerequisite workspace (POST /api/v1/curricula) operates on
the real catalog end-to-end, including a cross-subject prerequisite --
the "set up subject prerequisites" capability requested for Phase 6.

Full backend suite: 524/524 passing. Pint and PHPStan level 8 clean.
Frontend typecheck clean after widening the units Zod schema.

+++ backend/app/Models/Enrollment.php
@@ -16,7 +16,7 @@ use Illuminate\Database\Eloquent\Relations\HasOne;
  * @property int $student_id
  * @property int $academic_term_id
  * @property EnrollmentStatus $status
- * @property int $total_units
+ * @property float $total_units
  * @property ?CarbonImmutable $submitted_at
  * @property ?CarbonImmutable $registrar_decided_at
  * @property ?CarbonImmutable $payment_confirmed_at
@@ -58,7 +58,7 @@ final class Enrollment extends Model
     {
         return [
             'status' => EnrollmentStatus::class,
-            'total_units' => 'integer',
+            'total_units' => 'float',
             'submitted_at' => 'immutable_datetime',
             'registrar_decided_at' => 'immutable_datetime',
             'payment_confirmed_at' => 'immutable_datetime',
